Splits `printer/util.ex` into several modules
This commit splits the `JS2E.Printer.Util` module into a whole `printer/utils`
folder in order to increase cohesion, i.e. have one util module per relevant
area of printing.
