+++ lib/printer/all_of_printer.ex
@@ -5,10 +5,28 @@ defmodule JS2E.Printer.AllOfPrinter do
   """
 
   require Elixir.{EEx, Logger}
-  alias JS2E.Printer.{ErrorUtil, Util, PrinterError, PrinterResult}
-  alias JS2E.Printer.Utils.{Naming, Indentation}
+  alias JS2E.Printer.{ErrorUtil, PrinterError, PrinterResult}
+
+  alias JS2E.Printer.Utils.{
+    Naming,
+    Indentation,
+    ElmTypes,
+    ElmDecoders,
+    ElmEncoders,
+    ResolveType,
+    CommonOperations
+  }
+
   alias JS2E.{TypePath, Types}
-  alias JS2E.Types.{AllOfType, ObjectType, SchemaDefinition}
+
+  alias JS2E.Types.{
+    AllOfType,
+    EnumType,
+    OneOfType,
+    ObjectType,
+    UnionType,
+    SchemaDefinition
+  }
 
   @templates_location Application.get_env(:js2e, :templates_location)
 
@@ -39,7 +57,7 @@ defmodule JS2E.Printer.AllOfPrinter do
     {type_fields, errors} =
       types
       |> create_type_fields(path, schema_def, schema_dict, module_name)
-      |> Util.split_ok_and_errors()
+      |> CommonOperations.split_ok_and_errors()
 
     type_name
     |> type_template(type_fields)
@@ -76,8 +94,8 @@ defmodule JS2E.Printer.AllOfPrinter do
        ) do
     field_type_result =
       type_path
-      |> Util.resolve_type(parent, schema_def, schema_dict)
-      |> Util.create_type_name(schema_def, module_name)
+      |> ResolveType.resolve_type(parent, schema_def, schema_dict)
+      |> ElmTypes.create_type_name(schema_def, module_name)
 
     case field_type_result do
       {:ok, field_type} ->
@@ -117,7 +135,7 @@ defmodule JS2E.Printer.AllOfPrinter do
     {clauses, errors} =
       type_paths
       |> create_decoder_clauses(path, schema_def, schema_dict, module_name)
-      |> Util.split_ok_and_errors()
+      |> CommonOperations.split_ok_and_errors()
 
     normalized_name = Naming.normalize_identifier(name, :downcase)
     decoder_name = "#{normalized_name}Decoder"
@@ -163,21 +181,24 @@ defmodule JS2E.Printer.AllOfPrinter do
          module_name
        ) do
     with {:ok, {property_type, resolved_schema_def}} <-
-           Util.resolve_type(type_path, parent, schema_def, schema_dict),
+           ResolveType.resolve_type(type_path, parent, schema_def, schema_dict),
          {:ok, decoder_name} <-
-           Util.create_decoder_name(
+           ElmDecoders.create_decoder_name(
              {:ok, {property_type, resolved_schema_def}},
              schema_def,
              module_name
            ) do
       property_name = property_type.name
 
-      cond do
-        Util.union_type?(property_type) or Util.one_of_type?(property_type) ->
+      case property_type do
+        %OneOfType{} ->
           create_decoder_union_clause(property_name, decoder_name)
 
-        Util.enum_type?(property_type) ->
-          case Util.determine_primitive_type_decoder(property_type.type) do
+        %UnionType{} ->
+          create_decoder_union_clause(property_name, decoder_name)
+
+        %EnumType{} ->
+          case ElmDecoders.determine_primitive_type_decoder(property_type.type) do
             {:ok, property_type_decoder} ->
               create_decoder_enum_clause(
                 property_name,
@@ -189,7 +210,7 @@ defmodule JS2E.Printer.AllOfPrinter do
               {:error, error}
           end
 
-        true ->
+        _ ->
           create_decoder_normal_clause(property_name, decoder_name)
       end
     else
@@ -253,7 +274,7 @@ defmodule JS2E.Printer.AllOfPrinter do
     {encoder_properties, errors} =
       type_paths
       |> create_encoder_properties(path, schema_def, schema_dict, module_name)
-      |> Util.split_ok_and_errors()
+      |> CommonOperations.split_ok_and_errors()
 
     argument_name = Naming.normalize_identifier(name, :downcase)
     type_name = Naming.upcase_first(argument_name)
@@ -280,7 +301,7 @@ defmodule JS2E.Printer.AllOfPrinter do
          module_name
        ) do
     type_paths
-    |> Enum.map(&Util.resolve_type(&1, parent, schema_def, schema_dict))
+    |> Enum.map(&ResolveType.resolve_type(&1, parent, schema_def, schema_dict))
     |> Enum.map(&to_encoder_property(&1, schema_def, module_name))
     |> Enum.concat()
   end
@@ -302,9 +323,14 @@ defmodule JS2E.Printer.AllOfPrinter do
 
     type_def.properties
     |> Enum.map(fn {_child_name, child_path} ->
-      case Util.resolve_type(child_path, type_def.path, schema_def, schema_dict) do
+      case ResolveType.resolve_type(
+             child_path,
+             type_def.path,
+             schema_def,
+             schema_dict
+           ) do
         {:ok, {child_type_def, child_schema_def}} ->
-          case Util.create_encoder_name(
+          case ElmEncoders.create_encoder_name(
                  {:ok, {child_type_def, child_schema_def}},
                  schema_dict,
                  module_name
