+++ lib/printer/any_of_printer.ex
@@ -5,10 +5,20 @@ defmodule JS2E.Printer.AnyOfPrinter do
   """
 
   require Elixir.{EEx, Logger}
-  alias JS2E.Printer.{Util, PrinterError, PrinterResult}
-  alias JS2E.Printer.Utils.{Naming, Indentation}
+  alias JS2E.Printer.{PrinterError, PrinterResult}
+
+  alias JS2E.Printer.Utils.{
+    Naming,
+    Indentation,
+    ElmTypes,
+    ElmDecoders,
+    ElmEncoders,
+    ResolveType,
+    CommonOperations
+  }
+
   alias JS2E.{TypePath, Types}
-  alias JS2E.Types.{AnyOfType, SchemaDefinition}
+  alias JS2E.Types.{AnyOfType, EnumType, OneOfType, UnionType, SchemaDefinition}
 
   @templates_location Application.get_env(:js2e, :templates_location)
 
@@ -38,7 +48,7 @@ defmodule JS2E.Printer.AnyOfPrinter do
     {type_fields, errors} =
       types
       |> create_type_fields(path, schema_def, schema_dict, module_name)
-      |> Util.split_ok_and_errors()
+      |> CommonOperations.split_ok_and_errors()
 
     type_name
     |> type_template(type_fields)
@@ -75,8 +85,8 @@ defmodule JS2E.Printer.AnyOfPrinter do
        ) do
     field_type_result =
       type_path
-      |> Util.resolve_type(parent, schema_def, schema_dict)
-      |> Util.create_type_name(schema_def, module_name)
+      |> ResolveType.resolve_type(parent, schema_def, schema_dict)
+      |> ElmTypes.create_type_name(schema_def, module_name)
 
     case field_type_result do
       {:ok, field_type} ->
@@ -113,7 +123,7 @@ defmodule JS2E.Printer.AnyOfPrinter do
     {decoder_clauses, errors} =
       type_paths
       |> create_decoder_clauses(path, schema_def, schema_dict, module_name)
-      |> Util.split_ok_and_errors()
+      |> CommonOperations.split_ok_and_errors()
 
     normalized_name = Naming.normalize_identifier(name, :upcase)
     decoder_name = "#{name}Decoder"
@@ -159,21 +169,18 @@ defmodule JS2E.Printer.AnyOfPrinter do
          module_name
        ) do
     with {:ok, {property_type, resolved_schema_def}} <-
-           Util.resolve_type(type_path, parent, schema_def, schema_dict),
+           ResolveType.resolve_type(type_path, parent, schema_def, schema_dict),
          {:ok, decoder_name} <-
-           Util.create_decoder_name(
+           ElmDecoders.create_decoder_name(
              {:ok, {property_type, resolved_schema_def}},
              schema_def,
              module_name
            ) do
       property_name = property_type.name
 
-      cond do
-        Util.union_type?(property_type) or Util.one_of_type?(property_type) ->
-          create_decoder_union_clause(property_name, decoder_name)
-
-        Util.enum_type?(property_type) ->
-          case Util.determine_primitive_type_decoder(property_type.type) do
+      case property_type do
+        %EnumType{} ->
+          case ElmDecoders.determine_primitive_type_decoder(property_type.type) do
             {:ok, property_type_decoder} ->
               create_decoder_enum_clause(
                 property_name,
@@ -185,7 +192,13 @@ defmodule JS2E.Printer.AnyOfPrinter do
               {:error, error}
           end
 
-        true ->
+        %OneOfType{} ->
+          create_decoder_union_clause(property_name, decoder_name)
+
+        %UnionType{} ->
+          create_decoder_union_clause(property_name, decoder_name)
+
+        _ ->
           create_decoder_normal_clause(property_name, decoder_name)
       end
     else
@@ -245,7 +258,7 @@ defmodule JS2E.Printer.AnyOfPrinter do
     {encoder_properties, errors} =
       type_paths
       |> create_encoder_properties(path, schema_def, schema_dict, module_name)
-      |> Util.split_ok_and_errors()
+      |> CommonOperations.split_ok_and_errors()
 
     argument_name = Naming.normalize_identifier(name, :downcase)
     type_name = Naming.upcase_first(argument_name)
@@ -272,7 +285,7 @@ defmodule JS2E.Printer.AnyOfPrinter do
          module_name
        ) do
     type_paths
-    |> Enum.map(&Util.resolve_type(&1, parent, schema_def, schema_dict))
+    |> Enum.map(&ResolveType.resolve_type(&1, parent, schema_def, schema_dict))
     |> Enum.map(&to_encoder_property(&1, schema_def, module_name))
   end
 
@@ -285,7 +298,7 @@ defmodule JS2E.Printer.AnyOfPrinter do
   defp to_encoder_property({:error, error}, _sf, _md), do: {:error, error}
 
   defp to_encoder_property({:ok, {property, schema}}, schema_def, module_name) do
-    case Util.create_encoder_name(
+    case ElmEncoders.create_encoder_name(
            {:ok, {property, schema}},
            schema_def,
            module_name
