+++ lib/printer/printer.ex
@@ -5,9 +5,44 @@ defmodule JS2E.Printer do
   """
 
   require Logger
-  alias JS2E.Printer.{Util, PreamblePrinter, PrinterResult, SchemaResult}
+
+  alias JS2E.Printer.{
+    ErrorUtil,
+    PreamblePrinter,
+    PrinterResult,
+    SchemaResult
+  }
+
   alias JS2E.Types
-  alias JS2E.Types.SchemaDefinition
+
+  alias JS2E.Types.{
+    AllOfType,
+    AnyOfType,
+    ArrayType,
+    EnumType,
+    ObjectType,
+    OneOfType,
+    PrimitiveType,
+    TupleType,
+    TypeReference,
+    UnionType,
+    SchemaDefinition
+  }
+
+  alias JS2E.Printer.{
+    ErrorUtil,
+    AllOfPrinter,
+    AnyOfPrinter,
+    ArrayPrinter,
+    EnumPrinter,
+    ObjectPrinter,
+    OneOfPrinter,
+    PrimitivePrinter,
+    TuplePrinter,
+    TypeReferencePrinter,
+    UnionPrinter,
+    PrinterResult
+  }
 
   @spec print_schemas(Types.schemaDictionary(), String.t()) :: SchemaResult
   def print_schemas(schema_dict, module_name \\ "") do
@@ -56,13 +91,7 @@ defmodule JS2E.Printer do
       |> Enum.sort(&(&1.name < &2.name))
 
     types_result =
-      merge_results(
-        values,
-        schema_def,
-        schema_dict,
-        module_name,
-        &Util.print_type/4
-      )
+      merge_results(values, schema_def, schema_dict, module_name, &print_type/4)
 
     decoders_result =
       merge_results(
@@ -70,7 +99,7 @@ defmodule JS2E.Printer do
         schema_def,
         schema_dict,
         module_name,
-        &Util.print_decoder/4
+        &print_decoder/4
       )
 
     encoders_result =
@@ -79,7 +108,7 @@ defmodule JS2E.Printer do
         schema_def,
         schema_dict,
         module_name,
-        &Util.print_encoder/4
+        &print_encoder/4
       )
 
     printer_result =
@@ -137,5 +166,255 @@ defmodule JS2E.Printer do
     |> Enum.reduce(PrinterResult.new(), fn type_result, acc ->
       PrinterResult.merge(acc, type_result)
     end)
+  end
+
+  @spec print_type(
+          Types.typeDefinition(),
+          SchemaDefinition.t(),
+          Types.schemaDictionary(),
+          String.t()
+        ) :: PrinterResult.t()
+  def print_type(type_def, schema_def, schema_dict, module_name) do
+    case type_def do
+      %AllOfType{} ->
+        AllOfPrinter.print_type(type_def, schema_def, schema_dict, module_name)
+
+      %AnyOfType{} ->
+        AnyOfPrinter.print_type(type_def, schema_def, schema_dict, module_name)
+
+      %ArrayType{} ->
+        ArrayPrinter.print_type(type_def, schema_def, schema_dict, module_name)
+
+      %EnumType{} ->
+        EnumPrinter.print_type(type_def, schema_def, schema_dict, module_name)
+
+      %ObjectType{} ->
+        ObjectPrinter.print_type(type_def, schema_def, schema_dict, module_name)
+
+      %OneOfType{} ->
+        OneOfPrinter.print_type(type_def, schema_def, schema_dict, module_name)
+
+      %PrimitiveType{} ->
+        PrimitivePrinter.print_type(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %TupleType{} ->
+        TuplePrinter.print_type(type_def, schema_def, schema_dict, module_name)
+
+      %TypeReference{} ->
+        TypeReferencePrinter.print_type(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %UnionType{} ->
+        UnionPrinter.print_type(type_def, schema_def, schema_dict, module_name)
+
+      _ ->
+        struct_name = get_struct_name(type_def)
+        PrinterResult.new("", [ErrorUtil.unknown_type(struct_name)])
+    end
+  end
+
+  @spec print_decoder(
+          Types.typeDefinition(),
+          SchemaDefinition.t(),
+          Types.schemaDictionary(),
+          String.t()
+        ) :: PrinterResult.t()
+  def print_decoder(type_def, schema_def, schema_dict, module_name) do
+    case type_def do
+      %AllOfType{} ->
+        AllOfPrinter.print_decoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %AnyOfType{} ->
+        AnyOfPrinter.print_decoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %ArrayType{} ->
+        ArrayPrinter.print_decoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %EnumType{} ->
+        EnumPrinter.print_decoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %ObjectType{} ->
+        ObjectPrinter.print_decoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %OneOfType{} ->
+        OneOfPrinter.print_decoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %PrimitiveType{} ->
+        PrimitivePrinter.print_decoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %TupleType{} ->
+        TuplePrinter.print_decoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %TypeReference{} ->
+        TypeReferencePrinter.print_decoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %UnionType{} ->
+        UnionPrinter.print_decoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      _ ->
+        struct_name = get_struct_name(type_def)
+        PrinterResult.new("", [ErrorUtil.unknown_type(struct_name)])
+    end
+  end
+
+  @spec print_encoder(
+          Types.typeDefinition(),
+          SchemaDefinition.t(),
+          Types.schemaDictionary(),
+          String.t()
+        ) :: PrinterResult.t()
+  def print_encoder(type_def, schema_def, schema_dict, module_name) do
+    case type_def do
+      %AllOfType{} ->
+        AllOfPrinter.print_encoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %AnyOfType{} ->
+        AnyOfPrinter.print_encoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %ArrayType{} ->
+        ArrayPrinter.print_encoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %EnumType{} ->
+        EnumPrinter.print_encoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %ObjectType{} ->
+        ObjectPrinter.print_encoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %OneOfType{} ->
+        OneOfPrinter.print_encoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %PrimitiveType{} ->
+        PrimitivePrinter.print_encoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %TupleType{} ->
+        TuplePrinter.print_encoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %TypeReference{} ->
+        TypeReferencePrinter.print_encoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      %UnionType{} ->
+        UnionPrinter.print_encoder(
+          type_def,
+          schema_def,
+          schema_dict,
+          module_name
+        )
+
+      _ ->
+        struct_name = get_struct_name(type_def)
+        PrinterResult.new("", [ErrorUtil.unknown_type(struct_name)])
+    end
+  end
+
+  @spec get_struct_name(struct) :: String.t()
+  defp get_struct_name(struct) do
+    struct.__struct__
+    |> to_string
+    |> String.split(".")
+    |> List.last()
   end
 end
