--- test/printer/util_test.exs
@@ -1,4 +1,0 @@
-defmodule JS2ETest.Printer.Util do
-  use ExUnit.Case, async: true
-  doctest JS2E.Printer.Util, import: true
-end
