+++ test/printer/utils/common_operations_test.exs
@@ -1,0 +1,4 @@
+defmodule JS2ETest.Printer.Utils.CommonOperations do
+  use ExUnit.Case, async: true
+  doctest JS2E.Printer.Utils.CommonOperations, import: true
+end
