+++ test/printer/utils/elm_decoders_test.exs
@@ -1,0 +1,4 @@
+defmodule JS2ETest.Printer.Utils.ElmDecoders do
+  use ExUnit.Case, async: true
+  doctest JS2E.Printer.Utils.ElmDecoders, import: true
+end
