+++ test/printer/utils/elm_encoders_test.exs
@@ -1,0 +1,4 @@
+defmodule JS2ETest.Printer.Utils.ElmEncoders do
+  use ExUnit.Case, async: true
+  doctest JS2E.Printer.Utils.ElmEncoders, import: true
+end
