+++ test/printer/utils/elm_types_test.exs
@@ -1,0 +1,4 @@
+defmodule JS2ETest.Printer.Utils.ElmTypes do
+  use ExUnit.Case, async: true
+  doctest JS2E.Printer.Utils.ElmTypes, import: true
+end
