+++ test/printer/utils/indentation_test.exs
@@ -1,0 +1,4 @@
+defmodule JS2ETest.Printer.Utils.Indentation do
+  use ExUnit.Case, async: true
+  doctest JS2E.Printer.Utils.Indentation, import: true
+end
