+++ test/printer/utils/naming_test.exs
@@ -1,0 +1,4 @@
+defmodule JS2ETest.Printer.Utils.Naming do
+  use ExUnit.Case, async: true
+  doctest JS2E.Printer.Utils.Naming, import: true
+end
