+++ test/printer/utils/resolve_type_test.exs
@@ -1,0 +1,4 @@
+defmodule JS2ETest.Printer.Utils.ResolveType do
+  use ExUnit.Case, async: true
+  doctest JS2E.Printer.Utils.ResolveType, import: true
+end
